chore(project): reorganize project structure, add scripts for compiling main.exe file; docs(task\Architecture and structure.md): update docs
