+++ task/Architecture and structure.md
@@ -1,61 +1,109 @@
-> [!NOTE] pathes are the file relative
+### Structure
+
+./task
+├── Architecture and structure.md
+├── controllers
+│ ├── change_task_delay.c
+│ ├── get_callback.c
+│ ├── register_task.c
+│ └── remove_task.c
+├── environment
+│ ├── arguments.c
+│ ├── arguments.h
+│ ├── arguments_config.h
+│ ├── config.h
+│ ├── global_variables.c
+│ └── global_variables.h
+├── main.c
+├── model
+│ ├── change_task_delay_config.h
+│ ├── get_callback_config.h
+│ ├── handle_change_task_delay.c
+│ ├── handle_events_tasks.c
+│ ├── handle_events_tasks.h
+│ ├── handle_events_tasks_config.h
+│ ├── handle_get_callback.c
+│ ├── handle_register_task.c
+│ ├── handle_remove_task.c
+│ ├── register_task_config.h
+│ └── remove_task_config.h
+├── module_run_tasks_after_delay.h
+├── show_task_info.c
+├── technical specification.md
+├── tests
+│ ├── handle_id.test.c
+│ └── main.tests.c
+└── utilities
+├── handle_id.c
+├── handle_id_config.h
+├── sort_tasks_descending_by_delay_func.c
+└── utils.h
 
 ---
 
 #### Environment
 
-./config.h
+config.h
 
-./global_variables.h  
-./global_variables.c
+global_variables.h  
+global_variables.c
 
-./arguments.h
-./arguments_config.h
-./arguments.c
+arguments.h
+arguments_config.h
+arguments.c
 
 #### Model layer
 
-./handle_events_tasks.h  
-./handle_events_tasks_config.h  
-./handle_events_tasks.c
+handle_events_tasks.h  
+handle_events_tasks_config.h  
+handle_events_tasks.c
 
-./register_task_config.h  
-./handle_register_task.c
+register_task_config.h  
+handle_register_task.c
 
-./get_callback_config.h  
-./handle_get_callback.c
+get_callback_config.h  
+handle_get_callback.c
 
-./change_task_delay_config.h  
-./handle_change_task_delay.c
+change_task_delay_config.h  
+handle_change_task_delay.c
 
-./remove_task_config.h  
-./handle_remove_task.c
+remove_task_config.h  
+handle_remove_task.c
 
 #### Utilities
 
-./utils.h  
-./sort_tasks_descending_by_delay_func.c  
-./handle_id_config.h  
-./handle_id.c
+utils.h  
+sort_tasks_descending_by_delay_func.c  
+handle_id_config.h  
+handle_id.c
 
 #### Methods to use as module one (i.e. like a lib)
 
-./module_run_tasks_after_delay.h
+module_run_tasks_after_delay.h
 
 ---
 
-> [!NOTE] they're contoller layer
+#### Controllers
 
-./register_task.c  
-./get_callback.c  
-./change_task_delay.c  
-./remove_task.c
+register_task.c  
+get_callback.c  
+change_task_delay.c  
+remove_task.c
 
 ---
 
-> [!NOTE] example callback to use
+> [!NOTE] example callback to use, cwd
 
-./show_task_info.c
+show_task_info.c
+
+---
+
+### Tests
+
+> [!NOTE] partially manual tests scenarios
+
+handle_id.test.c
+main.tests.c
 
 ---
