+++ task/controllers/remove_task.c
@@ -1,5 +1,5 @@
-#include "./arguments.h"
-#include "./module_run_tasks_after_delay.h"
+#include "../environment/arguments.h"
+#include "../module_run_tasks_after_delay.h"
 
 extern bool is_remove_task;
