+++ task/environment/global_variables.c
