+++ task/environment/global_variables.h
