+++ task/main.c
@@ -1,4 +1,4 @@
-#include "./global_variables.h"
+#include "./environment/global_variables.h"
 #include "./module_run_tasks_after_delay.h"
 
 int main(void) {
@@ -7,7 +7,6 @@ int main(void) {
     // TODO! Fix this to desired logic from the ./Architecture and structure.md
     // !
     handle_events_tasks();
-
     usleep(100'000);
   }
 }
