+++ task/model/change_task_delay_config.h
@@ -1,7 +1,7 @@
 #ifndef CHANGE_TASK_DELAY_CONFIG_H
 #define CHANGE_TASK_DELAY_CONFIG_H
 
-#include "./config.h"
+#include "../environment/config.h"
 
 /**
  *  @details
