+++ task/module_run_tasks_after_delay.h
@@ -1,12 +1,12 @@
 #ifndef MODULE_RUN_TASKS_AFTER_DELAY_H
 #define MODULE_RUN_TASKS_AFTER_DELAY_H
 
-#include "./change_task_delay_config.h"
-#include "./config.h"
-#include "./get_callback_config.h"
-#include "./handle_events_tasks_config.h"
-#include "./register_task_config.h"
-#include "./remove_task_config.h"
+#include "./environment/config.h"
+#include "./model/change_task_delay_config.h"
+#include "./model/get_callback_config.h"
+#include "./model/handle_events_tasks_config.h"
+#include "./model/register_task_config.h"
+#include "./model/remove_task_config.h"
 
 PROMISE_HANDLE_EVENTS_TASKS handle_events_tasks(void);
 PROMISE_TASK_ID register_task(task_callback func_to_call, unsigned short arg,
