+++ task/show_task_info.c
@@ -1,4 +1,4 @@
-#include "./global_variables.h"
+#include "./environment/global_variables.h"
 #include "./module_run_tasks_after_delay.h"
 
 /**
