+++ task/tempCodeRunnerFile.sh
@@ -1,0 +1,1 @@
+show_task_info.c
