+++ task/utilities/handle_id_config.h
@@ -1,7 +1,7 @@
 #ifndef HANDLE_ID_H
 #define HANDLE_ID_H
 
-#include "./config.h"
+#include "../environment/config.h"
 
 /**
  *  @details
