+++ task/utilities/sort_tasks_descending_by_delay_func.c
