+++ task/utilities/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-#include "./config.h"
+#include "../environment/config.h"
 #include "./handle_id_config.h"
 
 void sort_tasks_descending_by_delay(Task arr[], size_t arr_size,
